# Use generic math for floating point formatting