--- a/src/libraries/System.Private.CoreLib/src/System/Half.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Half.cs
@@ -505,7 +505,7 @@ namespace System
         /// </summary>
         public override string ToString()
         {
-            return Number.FormatHalf(this, null, NumberFormatInfo.CurrentInfo);
+            return Number.FormatFloat(this, null, NumberFormatInfo.CurrentInfo);
         }
 
         /// <summary>
@@ -513,7 +513,7 @@ namespace System
         /// </summary>
         public string ToString([StringSyntax(StringSyntaxAttribute.NumericFormat)] string? format)
         {
-            return Number.FormatHalf(this, format, NumberFormatInfo.CurrentInfo);
+            return Number.FormatFloat(this, format, NumberFormatInfo.CurrentInfo);
         }
 
         /// <summary>
@@ -521,7 +521,7 @@ namespace System
         /// </summary>
         public string ToString(IFormatProvider? provider)
         {
-            return Number.FormatHalf(this, null, NumberFormatInfo.GetInstance(provider));
+            return Number.FormatFloat(this, null, NumberFormatInfo.GetInstance(provider));
         }
 
         /// <summary>
@@ -529,7 +529,7 @@ namespace System
         /// </summary>
         public string ToString([StringSyntax(StringSyntaxAttribute.NumericFormat)] string? format, IFormatProvider? provider)
         {
-            return Number.FormatHalf(this, format, NumberFormatInfo.GetInstance(provider));
+            return Number.FormatFloat(this, format, NumberFormatInfo.GetInstance(provider));
         }
 
         /// <summary>
@@ -542,13 +542,13 @@ namespace System
         /// <returns></returns>
         public bool TryFormat(Span<char> destination, out int charsWritten, [StringSyntax(StringSyntaxAttribute.NumericFormat)] ReadOnlySpan<char> format = default, IFormatProvider? provider = null)
         {
-            return Number.TryFormatHalf(this, format, NumberFormatInfo.GetInstance(provider), destination, out charsWritten);
+            return Number.TryFormatFloat(this, format, NumberFormatInfo.GetInstance(provider), destination, out charsWritten);
         }
 
         /// <inheritdoc cref="IUtf8SpanFormattable.TryFormat" />
         public bool TryFormat(Span<byte> utf8Destination, out int bytesWritten, [StringSyntax(StringSyntaxAttribute.NumericFormat)] ReadOnlySpan<char> format = default, IFormatProvider? provider = null)
         {
-            return Number.TryFormatHalf(this, format, NumberFormatInfo.GetInstance(provider), utf8Destination, out bytesWritten);
+            return Number.TryFormatFloat(this, format, NumberFormatInfo.GetInstance(provider), utf8Destination, out bytesWritten);
         }
 
         //