--- a/src/libraries/System.Private.CoreLib/src/System/Number.DiyFp.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Number.DiyFp.cs
@@ -68,6 +68,21 @@ namespace System
                 return result;
             }
 
+            // Computes the two boundaries of value.
+            //
+            // The bigger boundary (mPlus) is normalized.
+            // The lower boundary has the same exponent as mPlus.
+            //
+            // Precondition:
+            //  The value encoded by value must be greater than 0.
+            public static DiyFp CreateAndGetBoundaries<TNumber>(TNumber value, out DiyFp mMinus, out DiyFp mPlus)
+                where TNumber : unmanaged, IBinaryFloatParseAndFormatInfo<TNumber>
+            {
+                var result = Create(value);
+                result.GetBoundaries(TNumber.DenormalMantissaBits, out mMinus, out mPlus);
+                return result;
+            }
+
             public DiyFp(double value)
             {
                 Debug.Assert(double.IsFinite(value));
@@ -87,6 +102,15 @@ namespace System
                 Debug.Assert(Half.IsFinite(value));
                 Debug.Assert((float)value > 0.0f);
                 f = ExtractFractionAndBiasedExponent(value, out e);
+            }
+
+            public static DiyFp Create<TNumber>(TNumber value)
+                where TNumber : unmanaged, IBinaryFloatParseAndFormatInfo<TNumber>
+            {
+                Debug.Assert(TNumber.IsFinite(value));
+                Debug.Assert(value > TNumber.Zero);
+                ulong f = ExtractFractionAndBiasedExponent(value, out int e);
+                return new DiyFp(f, e);
             }
 
             public DiyFp(ulong f, int e)