--- a/src/libraries/System.Private.CoreLib/src/System/Number.Formatting.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Number.Formatting.cs
@@ -2653,7 +2653,7 @@ namespace System
                 // So f = (2^TrailingSignificandLength + mantissa), e = exp - ExponentBias - TrailingSignificandLength;
 
                 fraction |= (1UL << TNumber.DenormalMantissaBits);
-                exponent -= TNumber.ExponentBias - TNumber.DenormalMantissaBits;
+                exponent -= TNumber.ExponentBias + TNumber.DenormalMantissaBits;
             }
             else
             {