Extracted from #98643