@tannergooding 2024-05-29
Can you do some performance testing on Mono? The general process for doing so is documented here: https://github.com/dotnet/performance/blob/main/docs/benchmarking-workflow-dotnet-runtime.md

CC. @SamMonoRT