@huoyaoyuan 2024-05-30
@EgorBot -arm64 -intel -mono

```csharp
using System.Collections.Generic;
using System.Linq;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

namespace BenchmarkGround
{
    public class Program
    {
        public IEnumerable<object> Values => [0.0, 1.0, 1234.5678, 1.09876543211234567890e30];

        public IEnumerable<object> Formats => ["G", "F0", "F10", "E"];

        public IEnumerable<object[]> ValueAndFormats => Values.SelectMany(v => Formats.Select(f => new object[] { v, f }));

        [Benchmark]
        [ArgumentsSource(nameof(Values))]
        public string DefaultToString(double value) => value.ToString();

        [Benchmark]
        [ArgumentsSource(nameof(ValueAndFormats))]
        public string ToStringFormat(double value, string format) => value.ToString(format);
    }
}
```