@EgorBo 2024-05-30
@EgorBot -intel -mono --envvars MONO_ENV_OPTIONS:"--interpreter"

```csharp
using System.Collections.Generic;
using System.Linq;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

namespace BenchmarkGround
{
    public class Program
    {
        public IEnumerable<object> Values => [0.0, 1.0, 1234.5678, 1.09876543211234567890e30];

        public IEnumerable<object> Formats => ["G", "F0", "F10", "E"];

        public IEnumerable<object[]> ValueAndFormats => Values.SelectMany(v => Formats.Select(f => new object[] { v, f }));

        [Benchmark]
        [ArgumentsSource(nameof(Values))]
        public string DefaultToString(double value) => value.ToString();

        [Benchmark]
        [ArgumentsSource(nameof(ValueAndFormats))]
        public string ToStringFormat(double value, string format) => value.ToString(format);
    }
}
```