@EgorBo 2024-05-30
Yeah looks like Mono-JIT and Mono interp are fine with the changes in this PR, at least for the provided benchmark. Unfortunately the bot doesn't yet support LLVM-AOT to check that as well.