@fanyang-mono 2024-05-30
There are instructions here to run microbenchmarks with Mono AOT. (https://github.com/dotnet/performance/blob/main/docs/benchmarking-workflow-dotnet-runtime.md#run-the-benchmarks-with-aot)

Please test it before merge. Usually, AOT takes the most hit from library changes.