@huoyaoyuan 2024-06-01
The instruction I used for benchmark:
```
dotnet run -c Release -f net8.0 --affinity 65535  --runtimes monoaotllvm  --aotcompilerpath "C:\Users\Meow\source\dotnet\BenchmarkTestHost\monoAOT_PR\sgen\mini\mono-sgen.exe" --customruntimepack "C:\Users\Meow\source\dotnet\BenchmarkTestHost\monoAOT_PR\pack"  --aotcompilermode llvm
```

Results:

PR:
```
BenchmarkDotNet v0.13.12, Windows 11 (10.0.22631.3593/23H2/2023Update/SunValley3)
13th Gen Intel Core i9-13900K, 1 CPU, 32 logical and 24 physical cores
.NET SDK 9.0.100-preview.4.24267.66
  [Host]     : .NET 8.0.5 (8.0.524.21615), X64 RyuJIT AVX2
  Job-FQWQJE : .NET 9.0.0 (42.42.42.42424) using MonoVM, X64 X86Base

Affinity=00000000000000001111111111111111  Runtime=MonoAOTLLVM  Toolchain=MonoAOTLLVM
```

| Method          | value                | format | Mean        | Error    | StdDev   |
|---------------- |--------------------- |------- |------------:|---------:|---------:|
| ToStringFormat  | 0                    | E      |   126.13 ns | 0.378 ns | 0.354 ns |
| ToStringFormat  | 0                    | F0     |    71.27 ns | 0.440 ns | 0.412 ns |
| ToStringFormat  | 0                    | F10    |   104.93 ns | 0.545 ns | 0.510 ns |
| ToStringFormat  | 0                    | G      |    59.68 ns | 0.524 ns | 0.490 ns |
| DefaultToString | 0                    | ?      |    54.44 ns | 0.567 ns | 0.530 ns |
| ToStringFormat  | 1                    | E      |   359.80 ns | 1.770 ns | 1.655 ns |
| ToStringFormat  | 1                    | F0     |   258.08 ns | 1.218 ns | 1.080 ns |
| ToStringFormat  | 1                    | F10    |   298.01 ns | 1.273 ns | 1.191 ns |
| ToStringFormat  | 1                    | G      |   154.68 ns | 0.510 ns | 0.477 ns |
| DefaultToString | 1                    | ?      |   148.94 ns | 0.666 ns | 0.623 ns |
| ToStringFormat  | 1234.5678            | E      |   214.97 ns | 1.181 ns | 1.105 ns |
| ToStringFormat  | 1234.5678            | F0     |   338.89 ns | 1.577 ns | 1.475 ns |
| ToStringFormat  | 1234.5678            | F10    |   501.09 ns | 2.710 ns | 2.535 ns |
| ToStringFormat  | 1234.5678            | G      |   212.78 ns | 1.048 ns | 0.929 ns |
| DefaultToString | 1234.5678            | ?      |   209.32 ns | 1.529 ns | 1.430 ns |
| ToStringFormat  | 1.098(...)6E+30 [22] | E      |   196.33 ns | 1.254 ns | 1.112 ns |
| ToStringFormat  | 1.098(...)6E+30 [22] | F0     | 1,053.69 ns | 5.062 ns | 4.487 ns |
| ToStringFormat  | 1.098(...)6E+30 [22] | F10    | 1,189.30 ns | 5.812 ns | 5.436 ns |
| ToStringFormat  | 1.098(...)6E+30 [22] | G      |   287.90 ns | 2.102 ns | 1.966 ns |
| DefaultToString | 1.098(...)6E+30 [22] | ?      |   282.25 ns | 1.599 ns | 1.496 ns |

main:

| Method          | value                | format | Mean        | Error     | StdDev    |
|---------------- |--------------------- |------- |------------:|----------:|----------:|
| ToStringFormat  | 0                    | E      |   120.53 ns |  1.899 ns |  1.776 ns |
| ToStringFormat  | 0                    | F0     |    70.56 ns |  0.576 ns |  0.481 ns |
| ToStringFormat  | 0                    | F10    |   103.54 ns |  1.347 ns |  1.260 ns |
| ToStringFormat  | 0                    | G      |    57.83 ns |  0.352 ns |  0.329 ns |
| DefaultToString | 0                    | ?      |    54.52 ns |  0.466 ns |  0.413 ns |
| ToStringFormat  | 1                    | E      |   345.41 ns |  5.116 ns |  4.786 ns |
| ToStringFormat  | 1                    | F0     |   260.28 ns |  0.619 ns |  0.549 ns |
| ToStringFormat  | 1                    | F10    |   305.23 ns |  3.058 ns |  2.860 ns |
| ToStringFormat  | 1                    | G      |   149.78 ns |  1.623 ns |  1.518 ns |
| DefaultToString | 1                    | ?      |   146.37 ns |  1.783 ns |  1.668 ns |
| ToStringFormat  | 1234.5678            | E      |   220.22 ns |  2.379 ns |  2.109 ns |
| ToStringFormat  | 1234.5678            | F0     |   351.67 ns |  2.425 ns |  2.150 ns |
| ToStringFormat  | 1234.5678            | F10    |   512.52 ns |  8.060 ns |  7.539 ns |
| ToStringFormat  | 1234.5678            | G      |   208.87 ns |  0.888 ns |  0.831 ns |
| DefaultToString | 1234.5678            | ?      |   204.73 ns |  0.748 ns |  0.699 ns |
| ToStringFormat  | 1.098(...)6E+30 [22] | E      |   194.07 ns |  2.028 ns |  1.798 ns |
| ToStringFormat  | 1.098(...)6E+30 [22] | F0     | 1,090.41 ns | 13.084 ns | 10.925 ns |
| ToStringFormat  | 1.098(...)6E+30 [22] | F10    | 1,208.15 ns | 14.228 ns | 13.309 ns |
| ToStringFormat  | 1.098(...)6E+30 [22] | G      |   290.50 ns |  2.323 ns |  2.173 ns |
| DefaultToString | 1.098(...)6E+30 [22] | ?      |   286.22 ns |  2.756 ns |  2.578 ns |

Looks no significant regression.